+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [1.1.0] - 2021-03-29
+
+* RxFuture: Added **.value** for add a new Future;
+```dart
+final rxFuture = RxFuture.of(myFuture);
+...
+
+rxFuture.value = newFuture;
+```
 ## [1.0.0] - 2021-03-03
 
 * Added Collections (RxList, RxSet, RxMap);